# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for the DevOps challenge.",
        "title": "DevOps Challenge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/deals": {
            "get": {
                "description": "Returns all deals from Pipedrive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deals"
                ],
                "summary": "Get all deals",
                "parameters": [
                    {
                        "type": "string",
                        "example": "cursor_token",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "custom_field_1,custom_field_2",
                        "name": "custom_fields",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 123,
                        "name": "filter_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1,2,3",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "next_activity_id,last_activity_id",
                        "name": "include_fields",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 100,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 101,
                        "name": "org_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 456,
                        "name": "owner_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 789,
                        "name": "person_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 202,
                        "name": "pipeline_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "update_time",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "desc",
                        "name": "sort_direction",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 303,
                        "name": "stage_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "open,winner",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2025-01-01T10:20:00Z",
                        "name": "updated_since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2025-01-02T10:20:00Z",
                        "name": "updated_until",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetDealsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new deal in Pipedrive using query parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deals"
                ],
                "summary": "Add a new deal",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2021-01-01T00:00:00Z",
                        "name": "add_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2021-01-01T00:00:00Z",
                        "name": "close_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "USD",
                        "name": "currency",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2021-01-01",
                        "name": "expected_close_date",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": false,
                        "name": "is_deleted",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1,2,3",
                        "name": "label_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Lost Reason",
                        "name": "lost_reason",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2021-01-01T00:00:00Z",
                        "name": "lost_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "org_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "owner_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "person_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "pipeline_id",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "example": 90,
                        "name": "probability",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2021-01-01T00:00:00Z",
                        "name": "stage_change_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "stage_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "open",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Deal Title",
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2021-01-01T00:00:00Z",
                        "name": "update_time",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "example": 200.5,
                        "name": "value",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 7,
                        "name": "visible_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2021-01-01T00:00:00Z",
                        "name": "won_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CreateDealResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CreateDealResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "api.GetDealsResponse": {
            "type": "object",
            "properties": {
                "additional_data": {},
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}